+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Add `Clone` to WSS un/subscribe message types
 - Add `Debug`, `Clone`, `Copy` to `KrakenWSSClient`
 - Add `pub` to additional fields on structs in private and public WSS messages
+- Change `OpenOrder` WSS type to use `OrderStatus` enum
 
 ### v0.0.1